vfs_ceph: explicit cast to uint64_t upon failure of ceph_statfs
When a call to 'ceph_statfs' from with 'cephwrap_disk_free' returns
non-zero status do an explicit cast to uint64_t for the negative (-1)
value returned by 'status_code'.

+++ source3/modules/vfs_ceph.c
@@ -349,7 +349,7 @@ static uint64_t cephwrap_disk_free(struct vfs_handle_struct *handle,
 	ret = ceph_statfs(handle->data, smb_fname->base_name, &statvfs_buf);
 	if (ret < 0) {
 		DBG_DEBUG("[CEPH] ceph_statfs returned %d\n", ret);
-		return status_code(ret);
+		return (uint64_t)status_code(ret);
 	}
 	/*
 	 * Provide all the correct values.
